fix(reactions): measure real slide-box ratio instead of assuming 16:9
The slide box declares aspect-ratio 16:9, but as a flex child it gets
height-compressed in short/wide windows (measured up to ~4.3:1 on the
presenter view), which is when letterboxes appear — and stamps anchored
with a constant 16:9 assumption stopped tracking the slide content and
grew with the box while stretching.

All stamp surfaces now derive content fractions from the box's measured
ratio (shared useElementAspectRatio, ResizeObserver) combined with the
image's natural ratio; audience tap mapping uses the click-time rect.

Verified: reaction at content (50%, 10%) stays at that exact slide point
and keeps 3.4% content width while the presenter window is resized from
1800x750 (box 2.96:1) to 2300x700 (box 4.29:1).
