+++ src/widgets/presentation-layout/ui/viewer/SlideContainer.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useRef, useState } from "react";
 import type { SyntheticEvent } from "react";
 import styled, { css } from "styled-components";
 import { HighlightedSlideStyles } from "./SlideViewer.styles";
@@ -6,7 +6,9 @@ import {
   slideContentFractions,
   stampBoxStyle,
   imageNaturalRatio,
+  SLIDE_BOX_ASPECT,
 } from "@/shared/lib/slide-geometry";
+import { useElementAspectRatio } from "@/shared/lib/use-element-aspect-ratio";
 
 /**
  * 공용 슬라이드 컨테이너
@@ -26,8 +28,11 @@ const SlideContainer = ({
   const hasSrc = typeof src === "string" && src.length > 0;
 
   // 레터박스를 제외한 실제 슬라이드 콘텐츠 영역 — 스탬프 좌표(콘텐츠 기준 %)의 기준
+  // 박스는 flex 축소로 16:9 에서 벗어날 수 있으므로 실측 비율을 쓴다
+  const boxRef = useRef<HTMLDivElement | null>(null);
+  const boxRatio = useElementAspectRatio(boxRef);
   const [naturalRatio, setNaturalRatio] = useState<number | null>(null);
-  const contentFractions = slideContentFractions(naturalRatio);
+  const contentFractions = slideContentFractions(naturalRatio, boxRatio ?? SLIDE_BOX_ASPECT);
 
   const handleImageLoad = (e: SyntheticEvent<HTMLImageElement>) => {
     setNaturalRatio(imageNaturalRatio(e.currentTarget));
@@ -41,7 +46,7 @@ const SlideContainer = ({
   };
 
   return (
-    <SlideBox highlight={highlight} data-testid={testId}>
+    <SlideBox ref={boxRef} highlight={highlight} data-testid={testId}>
       {hasSrc ? (
         <img
           key={src}
